refactor: FavoriteHelpInformationViewModel -> SettingViewModel
