+++ android/app/src/main/java/com/on/turip/domain/userstorage/TuripDeviceIdentifier.kt
@@ -2,4 +2,8 @@ package com.on.turip.domain.userstorage
 
 data class TuripDeviceIdentifier(
     val fid: String,
-)
+) {
+    companion object {
+        val EMPTY: TuripDeviceIdentifier = TuripDeviceIdentifier(fid = "")
+    }
+}
