--- android/app/src/main/java/com/on/turip/ui/main/favorite/FavoriteHelpInformationViewModel.kt
@@ -1,47 +1,0 @@
-package com.on.turip.ui.main.favorite
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import androidx.lifecycle.viewModelScope
-import com.on.turip.domain.userstorage.TuripDeviceIdentifier
-import com.on.turip.domain.userstorage.repository.UserStorageRepository
-import com.on.turip.ui.main.favorite.model.HelpInformationModelType
-import dagger.hilt.android.lifecycle.HiltViewModel
-import kotlinx.coroutines.launch
-import timber.log.Timber
-import javax.inject.Inject
-
-@HiltViewModel
-class FavoriteHelpInformationViewModel @Inject constructor(
-    private val userStorageRepository: UserStorageRepository,
-) : ViewModel() {
-    private val _helpInformationItems: MutableLiveData<List<HelpInformationModelType>> =
-        MutableLiveData<List<HelpInformationModelType>>()
-    val helpInformationItems: LiveData<List<HelpInformationModelType>> get() = _helpInformationItems
-
-    private val _userId: MutableLiveData<TuripDeviceIdentifier> = MutableLiveData()
-    val userId: LiveData<TuripDeviceIdentifier> get() = _userId
-
-    init {
-        loadId()
-        loadHelpInformationItems()
-    }
-
-    private fun loadId() {
-        viewModelScope.launch {
-            userStorageRepository
-                .loadId()
-                .onSuccess { result ->
-                    _userId.value = result
-                }.onFailure {
-                    Timber.e("${it.message}")
-                }
-        }
-    }
-
-    private fun loadHelpInformationItems() {
-        _helpInformationItems.value =
-            listOf(HelpInformationModelType.INQUIRY, HelpInformationModelType.PRIVACY_POLICY)
-    }
-}
